Fix Windows Debug build

+++ libevm/EVMC.cpp
@@ -33,7 +33,7 @@ owning_bytes_ref EVMC::exec(u256& io_gas, ExtVMFace& _ext, const OnOpFunc& _onOp
     (void)int64max;
     assert(io_gas <= int64max);
     assert(_ext.envInfo().gasLimit() <= int64max);
-    assert(_ext.depth <= std::numeric_limits<int32_t>::max());
+    assert(_ext.depth <= static_cast<size_t>(std::numeric_limits<int32_t>::max()));
 
     auto gas = static_cast<int64_t>(io_gas);
     EVM::Result r = execute(_ext, gas);
